+++ delta/app/src/main/scala/ch/epfl/bluebrain/nexus/delta/wiring/QuotasModule.scala
@@ -8,7 +8,7 @@ import ch.epfl.bluebrain.nexus.delta.rdf.utils.JsonKeyOrdering
 import ch.epfl.bluebrain.nexus.delta.routes.QuotasRoutes
 import ch.epfl.bluebrain.nexus.delta.sdk._
 import ch.epfl.bluebrain.nexus.delta.sdk.model.BaseUri
-import ch.epfl.bluebrain.nexus.delta.sdk.model.projects.ProjectsConfig.QuotasConfig
+import ch.epfl.bluebrain.nexus.delta.sdk.model.quotas.QuotasConfig
 import ch.epfl.bluebrain.nexus.delta.service.quotas.QuotasImpl
 import izumi.distage.model.definition.{Id, ModuleDef}
 import monix.execution.Scheduler
