+++ delta/sdk/src/main/scala/ch/epfl/bluebrain/nexus/delta/sdk/model/projects/ProjectsConfig.scala
@@ -6,7 +6,8 @@ import ch.epfl.bluebrain.nexus.delta.rdf.IriOrBNode.Iri
 import ch.epfl.bluebrain.nexus.delta.sdk.cache.KeyValueStoreConfig
 import ch.epfl.bluebrain.nexus.delta.sdk.model.Label
 import ch.epfl.bluebrain.nexus.delta.sdk.model.permissions.Permission
-import ch.epfl.bluebrain.nexus.delta.sdk.model.projects.ProjectsConfig.{AutomaticProvisioningConfig, QuotasConfig}
+import ch.epfl.bluebrain.nexus.delta.sdk.model.projects.ProjectsConfig.{AutomaticProvisioningConfig}
+import ch.epfl.bluebrain.nexus.delta.sdk.model.quotas.QuotasConfig
 import ch.epfl.bluebrain.nexus.delta.sdk.model.search.PaginationConfig
 import ch.epfl.bluebrain.nexus.delta.sourcing.config.{AggregateConfig, SaveProgressConfig}
 import pureconfig.ConfigReader
@@ -116,34 +117,6 @@ object ProjectsConfig {
       enabledRealms,
       ProjectFields(Some(description), apiMappings, base, vocab)
     )
-  }
-
-  /**
-    * The configuration for quotas on projects
-    *
-    * @param resources maximum number of resources per project
-    * @param enabled   flag to enable or disable project quotas
-    * @param custom    custom quotas for certain projects
-    */
-  final case class QuotasConfig(resources: Int, enabled: Boolean, custom: Map[ProjectRef, Int])
-  object QuotasConfig {
-
-    @nowarn("cat=unused")
-    implicit final val quotasConfigReader: ConfigReader[QuotasConfig] = {
-
-      implicit val customMapReader: ConfigReader[Map[ProjectRef, Int]] = genericMapReader[ProjectRef, Int] { key =>
-        key.split("/").toList match {
-          case orgStr :: projectStr :: Nil =>
-            (Label(orgStr), Label(projectStr))
-              .mapN(ProjectRef(_, _))
-              .leftMap(err => CannotConvert(key, classOf[ProjectRef].getSimpleName, err.getMessage))
-          case _                           =>
-            Left(CannotConvert(key, classOf[ProjectRef].getSimpleName, "Wrong format"))
-        }
-      }
-
-      deriveReader[QuotasConfig]
-    }
   }
 
   implicit final val projectConfigReader: ConfigReader[ProjectsConfig] =
